fix(msdns): stop one huge-TTL NS record from killing the whole zone sync
get_records() cast TimeToLive.TotalSeconds through PowerShell [int]
(System.Int32, max 2147483647) — a record with a TTL above that
(unusual but legal) threw a terminating conversion error and aborted
the entire zone's pipeline, not just that record. The NS fix exposed
this: NS records previously never reached the TTL cast at all, since
$data stayed null for them before that fix landed.

Same bug class as 0031 (DHCP IAID overflowing a signed INTEGER) — same
treatment: [int] -> [int64] in the PS cast, cache_dns_records.ttl
widened to BIGINT (migration 0032).

Adds PS-script-text and huge-TTL-parsing coverage to
test_msdns_get_records.py.

+++ backend/app/models/cache.py
@@ -17,7 +17,7 @@ class CachedDNSRecord(Base):
     record_type = Column(String, nullable=False)
     value       = Column(String, nullable=False)
     zone        = Column(String, nullable=False, index=True)
-    ttl         = Column(Integer, default=3600)
+    ttl         = Column(BigInteger, default=3600)  # spec allows up to uint32; overflows a plain Integer
     source      = Column(String, nullable=False, index=True)
     synced_at   = Column(DateTime, nullable=False)
